refactor: take out the processing child process listener server
